finishing auth

+++ backend/routes/controllers/post.js
@@ -28,7 +28,7 @@ export const addPost = (req, res) => {
     imagePath: `${url}/images/${req.file.filename}`,
     creator: req.userData.userId,
   });
-  console.log(`user data: ${req.userData}`);
+
   post
     .save()
     .then((createdPost) => {
@@ -86,18 +86,25 @@ export const updatePost = (req, res) => {
     title: req.body.title,
     content: req.body.content,
     imagePath: imagePath,
+    creator: req.userData.userId,
   });
   console.log(post);
 
-  Post.updateOne({ _id: req.params.id }, post).then((result) => {
-    res.status(201).json({ message: "update successful" });
+  Post.updateOne(
+    { _id: req.params.id, creator: req.userData.userId },
+    post
+  ).then((result) => {
+    if (result.nModified > 0)
+      res.status(201).json({ message: "update successful" });
+    else res.status(401).json({ message: "Not authorized" });
   });
 };
 
 export const deletePost = (req, res) => {
   const { id } = req.params;
 
-  Post.deleteOne({ _id: id }).then((result) => {
-    console.log(result), res.status(201).json({ message: "deleted" });
+  Post.deleteOne({ _id: id, creator: req.userData.creator }).then((result) => {
+    if (result.n > 0) res.status(201).json({ message: "Delete successful" });
+    else res.status(401).json({ message: "Not authorized" });
   });
 };
